requirements.txt
albumentations==0.4.6
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
cloudpickle==1.6.0
cmake==3.18.2.post1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
entrypoints==0.3
future==0.18.2
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.1
networkx==2.5
notebook==6.1.4
numpy==1.19.2
onnx==1.7.0
onnxruntime==1.3.0
opencv-python==4.4.0.44
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
scikit-build==0.11.1
scikit-image==0.17.2
scipy==1.5.2
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
submitit==1.1.1
terminado==0.9.1
testpath==0.4.4
tifffile==2020.9.3
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.50.0
traitlets==4.3.3
typing-extensions==3.7.4.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.2.0
